chore: move ui to src/
