+++ src/ui/views/announcement.py
