+++ src/ui/views/define_word.py
